+++ src/main/kotlin/gregtechlite/gtlitecore/api/capability/MultipleNotifiableHandler.kt
@@ -1,7 +1,7 @@
 package gregtechlite.gtlitecore.api.capability
 
 import gregtech.api.capability.INotifiableHandler
-import gregtechlite.gtlitecore.api.mixins.MixinExtension
+import gregtechlite.gtlitecore.mixins.MixinExtension
 
 @MixinExtension
 interface MultipleNotifiableHandler
